// projects/lib/src/core/types/style-config.type.ts
export type StyleConfig = {
  /**Utilities for controlling how flex and grid items are positioned along a container's cross axis.
     * @see https://tailwindcss.com/docs/align-items
     */
  alignItem: AlignItem;
  /**Utilities for controlling the direction of flex items.
   * @see https://tailwindcss.com/docs/flex-direction
   */
  flexDirection: FlexDirection;
  /**Utilities for controlling how flex items grow.
   * @see https://tailwindcss.com/docs/flex-grow
   */
  flexGrow: FlexGrow;
  /**Utilities for controlling how flex items shrink.
   * @see https://tailwindcss.com/docs/flex-shrink
   */
  flexShrink: FlexShrink;
  /**Utilities for controlling how flex items wrap.
   * @see https://tailwindcss.com/docs/flex-wrap
   */
  flexWrap: FlexWrap;
  /**Utilities for controlling gutters between grid and flexbox items.
   * @see https://tailwindcss.com/docs/gap
   */
  gap: Gap;
  /**Utilities for specifying the columns in a grid layout.
   * @see https://tailwindcss.com/docs/grid-template-columns
   */
  gridCols: GridColumns;
  /**Utilities for controlling how flex and grid items are positioned along a container's main axis.
   * @see https://tailwindcss.com/docs/justify-content
   */
  justifyContent: JustifyContent;
  /**Utilities for controlling how content is justified and aligned at the same time.
   * @see https://tailwindcss.com/docs/place-content
   */
  placeContent: PlaceContent;
  /**Utilities for controlling how items are justified and aligned at the same time.
   * @see https://tailwindcss.com/docs/place-items
   */
  placeItems: PlaceItems;
  /**Utilities for controlling how an individual flex or grid item is positioned along its container's cross axis.
   * @see https://tailwindcss.com/docs/align-self
   */
  alignSelf: AlignSelf;
  /**Utilities for controlling how rows are positioned in multi-row flex and grid containers.
   * @see https://tailwindcss.com/docs/align-content
   */
  alignContent: AlignContent;
  animation: TwAnimation;
  transitionDuration: TransitionDuration;
  transitionProperty: TransitionProperty;
  transitionTimingFunction: TransitionTimingFunction;
  ringColor: RingColor;
  ringOpacity: RingOpacity;
  ringOffsetWidth: RingOffsetWidth;
  borderType: BorderType;
  outlineWidth: OutlineWidth;
  outlineOffset: OutlineOffset;
  outlineColor: OutlineColor;
  divideColor: DivideColor;
  divideWidth: DivideWidth;
  textOpacity: TextOpacity;
  textColor: TextColor;
  bgColor: BackgroundColor;
  bgOpacity: BackgroundOpacity;
  bgGradient: BackgroundGradient;
  borderColor: BorderColor;
  borderOpacity: BorderOpacity;
  shadowColor: BoxShadowColor;
  display: Display;
  position: Position;
  borderRadius: BorderRadius;
  ringWidth: RingWidth;
  borderWidth: BorderWidth;
  fontWeight: FontWeight;
  fontSize: FontSize;
  overflow: Overflow;
  padding: Padding;
  margin: Margin;
  size: Size;
  width: Width;
  maxWidth: MaxWidth;
  height: Height;
  maxHeight: MaxHeight;
  textWrap: TextWrap;
  userSelect: UserSelect;
  textAlign: TextAlign;
  zIndex: ZIndex;
  cursor: Cursor;
  lineHeight: LineHeight;
  blur: Blur;
  isolation: 'isolate' | 'isolation-auto';
  backdrop: Backdrop;
  dropShadow: DropShadow;
  boxShadow: BoxShadow;
  opacity: Opacity;
  pointerEvents: PointerEvents;
  transformOrigin: TransformOrigin;
  translate: TwTranslate;
  verticalAlign: VerticalAlign;
  // Media queries modifiers
  sm: Partial<ModifiedStyle<'sm', Omit<StyleConfig, MediaQueryModifier>>>;
  md: Partial<ModifiedStyle<'md', Omit<StyleConfig, MediaQueryModifier>>>;
  lg: Partial<ModifiedStyle<'lg', Omit<StyleConfig, MediaQueryModifier>>>;
  // State modifiers
  focus: Partial<ModifiedStyle<'focus', StyleConfig>>;
  focusVisible: Partial<ModifiedStyle<'focus-visible', StyleConfig>>;
  hover: Partial<ModifiedStyle<'hover', StyleConfig>>;
  active: Partial<ModifiedStyle<'active', StyleConfig>>;
  disabled: Partial<ModifiedStyle<'disabled', StyleConfig>>;
  checked: Partial<ModifiedStyle<'checked', StyleConfig>>;
  // Style modifiers
  child: Partial<ModifiedStyle<'*', StyleConfig>>;
  first: Partial<ModifiedStyle<'first', StyleConfig>>;
  last: Partial<ModifiedStyle<'last', StyleConfig>>;
  dark: Partial<ModifiedStyle<'dark', StyleConfig>>;
};

/**A type to define styles that apply on modifiers. Ex.: sm: - focus: - click: - dark: etc.*/
export type ModifiedStyle<Modifier extends string, Type extends ValueObject> = {
  [p1 in keyof Type]: `${Modifier}:${Extract<Type[p1], string>}` | {
    [p2 in keyof Type[p1]]: `${Modifier}:${string}:${string}`
  }
};

export type MediaQueryModifier = 'sm' | 'md' | 'lg';